Moved benchmarks out to a separate file
