+++ dense_matop_test.go
@@ -413,17 +413,6 @@ var repeatTests = []repeatTest{
 	},
 }
 
-func BenchmarkDenseRepeat(b *testing.B) {
-	for _, tst := range repeatTests {
-		tst := tst
-		b.Run(tst.name, func(b *testing.B) {
-			for n := 0; n < b.N; n++ {
-				tst.tensor.Repeat(tst.axis, tst.repeats...)
-			}
-		})
-	}
-}
-
 func TestDense_Repeat(t *testing.T) {
 	assert := assert.New(t)
